requirements.txt
Flask==3.0.0
mysql-connector-python
gunicorn
python-dotenv
qrcode[pil]
Pillow>=9.1.0
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.16
twilio>=9.0.0
APScheduler>=3.10.0
# TiDB: protocolo MySQL (mysql-connector o PyMySQL con SQLAlchemy)
